+++ logfire/_internal/auto_trace/__init__.py
@@ -25,7 +25,7 @@ def install_auto_tracing(
     See `Logfire.install_auto_tracing` for more information.
     """
     if isinstance(modules, Sequence):
-        modules = modules_func_from_sequence(modules)  # type: ignore
+        modules = modules_func_from_sequence(modules)  # pyright: ignore[reportUnknownArgumentType]
 
     if not callable(modules):
         raise TypeError('modules must be a list of strings or a callable')